Switch to per fragment request locking
This ensures that each time interval is only requested once, previous
implementation was not able to ensure this.
The main issue with previous impl, was during the cache write back, a
concurrent request may start and won't have the same first missing
fragment and would start requesting already downloaded fragments but
not yet in cache.

The only minor downside of this new approach is that two concurrent
requests may interleave lock some fragments and thus split a larger
request in smaller concurrent ones.

Signed-off-by: Alexis Jeandet <alexis.jeandet@member.fsf.org>